fix error convert data

+++ BaiThucHanhTuan2/src/com/phenikaa/Main.java
@@ -33,15 +33,15 @@ public class Main {
         System.out.print("\nPeek :"+StackLinkedList.peek());
 
         System.out.println("\n");
-        long []arr={1,2,3,4,5};
-        long []arrReverse = new long[arr.length];
+        Integer []arr={1,2,3,4,5};
+        Integer []arrReverse = new Integer[arr.length];
         System.out.print("\nDanh sach cho truoc : ");
         for(long x:arr){
             System.out.print(x+" ");
         }
 
         Stack stack= new Stack(arr.length);
-        for(int i=0;i<arr.length;i++){
+        for(Integer i=0;i<arr.length;i++){
             stack.push(arr[i]);
         }
